refactor comscore unit tests
