+++ src/app/containers/ComscoreAnalytics/Amp/index.test.jsx
@@ -1,25 +1,10 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import ContextWrap from '../testHelper';
-import ComscoreAnalytics from '..';
+import AmpComscore from '.';
 
 describe('Snapshots', () => {
   it('should render comscore script when on amp and toggle is enabled', () => {
-    const toggleState = {
-      comscoreAnalytics: {
-        enabled: true,
-      },
-    };
-    const { container } = render(
-      <ContextWrap
-        platform="amp"
-        pageType="article"
-        origin="bbc.com"
-        toggleState={toggleState}
-      >
-        <ComscoreAnalytics />
-      </ContextWrap>,
-    );
+    const { container } = render(<AmpComscore />);
 
     expect(container.firstChild).not.toBeNull();
     expect(container.firstChild).toMatchSnapshot();
