+++ src/app/containers/ComscoreAnalytics/Canonical/index.test.jsx
@@ -1,27 +1,21 @@
 import React from 'react';
 import { render, waitFor } from '@testing-library/react';
 import { shouldMatchSnapshot } from '@bbc/psammead-test-helpers';
-import ContextWrap from '../testHelper';
-import ComscoreAnalytics from '..';
+import { UserContext } from '#contexts/UserContext';
+import CanonicalComscore from '.';
 
 describe('Canonical Comscore Analytics ', () => {
-  const toggleState = {
-    comscoreAnalytics: {
-      enabled: true,
-    },
-  };
+  beforeEach(() => {
+    /* eslint-disable no-underscore-dangle */
+    window._comscore = [];
+  });
 
   describe('Assertions', () => {
     it('should render null when not on client', async () => {
       const { container } = render(
-        <ContextWrap
-          platform="canonical"
-          pageType="article"
-          origin="bbc.com"
-          toggleState={toggleState}
-        >
-          <ComscoreAnalytics />
-        </ContextWrap>,
+        <UserContext.Provider value={{ personalisationEnabled: false }}>
+          <CanonicalComscore />
+        </UserContext.Provider>,
       );
 
       expect(container).toBeEmptyDOMElement();
@@ -31,14 +25,9 @@ describe('Canonical Comscore Analytics ', () => {
       const noScriptContent = `<img src="https://sb.scorecardresearch.com/p?c1=2&c2=17986528&cs_ucfr=0&cv=2.0&cj=1" />`;
 
       render(
-        <ContextWrap
-          platform="canonical"
-          pageType="article"
-          origin="bbc.com"
-          toggleState={toggleState}
-        >
-          <ComscoreAnalytics />
-        </ContextWrap>,
+        <UserContext.Provider value={{ personalisationEnabled: false }}>
+          <CanonicalComscore />
+        </UserContext.Provider>,
       );
       await waitFor(() => {
         const noScriptEl = document.querySelector('noscript');
@@ -49,23 +38,46 @@ describe('Canonical Comscore Analytics ', () => {
         expect(scriptEl).toBeInTheDocument();
         expect(scriptEl).toHaveAttribute('async');
         expect(scriptEl.src).toEqual(
-          'http://localhost:7080/static/js/comscore-1.0.js',
+          'http://localhost:7080/static/js/comscore/main-1.0.js',
         );
+      });
+    });
+
+    it('should create window._comscore without personalisation', async () => {
+      render(
+        <UserContext.Provider value={{ personalisationEnabled: false }}>
+          <CanonicalComscore />
+        </UserContext.Provider>,
+      );
+
+      await waitFor(() => {
+        expect(window._comscore).toEqual([
+          { c1: '2', c2: '17986528', cs_ucfr: '0' },
+        ]);
+      });
+    });
+
+    it('should create window._comscore with personalisation', async () => {
+      render(
+        <UserContext.Provider value={{ personalisationEnabled: true }}>
+          <CanonicalComscore />
+        </UserContext.Provider>,
+      );
+
+      await waitFor(() => {
+        expect(window._comscore).toEqual([
+          { c1: '2', c2: '17986528', cs_ucfr: '1' },
+        ]);
       });
     });
   });
 
   describe('Snapshots', () => {
     shouldMatchSnapshot(
-      'should render comscore script when on canonical and toggle is enabled',
-      <ContextWrap
-        platform="canonical"
-        pageType="article"
-        origin="bbc.com"
-        toggleState={toggleState}
-      >
-        <ComscoreAnalytics />
-      </ContextWrap>,
+      'should render comscore script when on canonical',
+      <UserContext.Provider value={{ personalisationEnabled: false }}>
+        <CanonicalComscore />
+      </UserContext.Provider>,
     );
   });
 });
