+++ src/app/containers/ComscoreAnalytics/index.test.jsx
@@ -1,7 +1,66 @@
 import React from 'react';
+import { node, string, shape } from 'prop-types';
 import { render } from '@testing-library/react';
-import ContextWrap from './testHelper';
+import { ServiceContextProvider } from '#contexts/ServiceContext';
+import { RequestContextProvider } from '#contexts/RequestContext';
+import { ToggleContext } from '#contexts/ToggleContext';
+import { UserContext } from '#contexts/UserContext';
 import ComscoreAnalytics from '.';
+
+const defaultToggleState = {
+  comscoreAnalytics: {
+    enabled: false,
+  },
+};
+
+const mockToggleDispatch = jest.fn();
+
+const defaultPersonalisation = { personalisationEnabled: false };
+
+const ContextWrap = ({
+  pageType,
+  platform,
+  origin,
+  children,
+  toggleState,
+  personalisation,
+}) => (
+  <RequestContextProvider
+    isAmp={platform === 'amp'}
+    pageType={pageType}
+    service="news"
+    statusCode={200}
+    bbcOrigin={origin}
+    pathname="/pathname"
+  >
+    <ServiceContextProvider service="pidgin">
+      <ToggleContext.Provider
+        value={{
+          toggleState,
+          toggleDispatch: mockToggleDispatch,
+        }}
+      >
+        <UserContext.Provider value={personalisation}>
+          {children}
+        </UserContext.Provider>
+      </ToggleContext.Provider>
+    </ServiceContextProvider>
+  </RequestContextProvider>
+);
+
+ContextWrap.propTypes = {
+  children: node.isRequired,
+  pageType: string.isRequired,
+  origin: string.isRequired,
+  platform: string.isRequired,
+  toggleState: shape({}),
+  personalisation: shape({}),
+};
+
+ContextWrap.defaultProps = {
+  toggleState: defaultToggleState,
+  personalisation: defaultPersonalisation,
+};
 
 describe('Comscore Analytics Container', () => {
   describe('Assertions - AMP', () => {
